Conflicts:
	src/oofemlib/domain.C
	src/oofemlib/mixedgradientpressureweakperiodic.C
	src/oofemlib/vtkxmlexportmodule.h
	src/oofemlib/xfem/enrichmentdomain.h
	src/oofemlib/xfem/enrichmentitem.C
	src/oofemlib/xfem/enrichmentitem.h
	src/oofemlib/xfem/xfemmanager.C
	src/sm/shell7base.h
	src/sm/tr2shell7xfem.C

+++ src/oofemlib/nrsolver.C
@@ -118,7 +118,7 @@ NRSolver :: initializeFrom(InputRecord *ir)
     IRResultType result;                // Required by IR_GIVE_FIELD macro
 
     // Choosing a big "enough" number. (Alternative: Force input of maxinter)
-    nsmax = 1e8;
+    nsmax = (int)1e8;
     IR_GIVE_OPTIONAL_FIELD(ir, nsmax, _IFT_NRSolver_maxiter);
     if ( nsmax < 0 ) {
         OOFEM_ERROR("NRSolver :: initializeFrom: nsmax < 0");
